fix: DataLoader collate pour PIL; dtype au lieu de torch_dtype (transformers récents)

+++ lora_3dsr_mindcube_workbench/infer_mindcube_llava.py
@@ -44,7 +44,7 @@ def main() -> None:
     t0 = time.perf_counter()
     processor = AutoProcessor.from_pretrained(args.base_model_id, trust_remote_code=True)
     base = LlavaNextForConditionalGeneration.from_pretrained(
-        args.base_model_id, torch_dtype=dtype, trust_remote_code=True
+        args.base_model_id, dtype=dtype, trust_remote_code=True
     )
     base = base.to("cuda")
     model = PeftModel.from_pretrained(base, args.adapter_dir)
